add authorization groups: admin, staff, user. add auth check for user resource and order resource

+++ rest_api/resources/order.py
@@ -1,16 +1,10 @@
 from flask_restful import Resource, reqparse
 from rest_api.models.order import OrderModel
-# from werkzeug.security import safe_str_cmp
-# from flask_jwt_extended import (
-#     create_access_token,
-#     create_refresh_token,
-#     jwt_refresh_token_required,
-#     get_jwt_identity,
-#     get_raw_jwt,
-#     jwt_required
-# )
-# from blacklist import BLACKLIST
-
+from flask_jwt_extended import (
+    jwt_required,
+    jwt_optional,
+    get_jwt_identity
+)
 
 ########################################
 #### allow staff to post new order #####
@@ -27,11 +21,19 @@ class OrderCreate(Resource):
         "staff_id", type = str, required=True, help="staff_id cannot be blank."
     )
 
-    # @jwt_required
+
+    @jwt_required
     def post(self):
         data = self.order_parser.parse_args()
 
         order = OrderModel.find_by_ur_code(data["ur_code"])
+
+        identity = get_jwt_identity()
+
+        # only admin and staff members are allowed to post new orders.
+        if identity["auth_level"] == "user":
+            return {"message":"unauthorized access, user cannot create order."},500
+
 
         if order:
             return {
@@ -57,7 +59,7 @@ class OrderInfo(Resource):
     order_parser.add_argument(
         "ur_code", type = str, required=True, help="ur_code cannot be blank."
     ) 
-    # @jwt_optional
+    @jwt_optional
     def post(self):
         data = self.order_parser.parse_args()
 
@@ -67,7 +69,10 @@ class OrderInfo(Resource):
             return {"message":"order with ur_code{} doen not exist.".format(data['ur_code'])},404
         
         # TODO 
-        # if with partial permission
+        # if staff:
+
+        # TODO
+        # if user:
 
 
         # if with full permission
@@ -94,7 +99,7 @@ class OrderUpdate(Resource):
     order_parser.add_argument(
         "user_id", type = str, required=True, help="staff_id cannot be blank."
     )
-    # @jwt_required
+    @jwt_required
     def put(self):
         data = self.order_parser.parse_args()
         
@@ -103,11 +108,20 @@ class OrderUpdate(Resource):
         if not order:
             return {"message":"order with ur_code{} doen not exist.".format(data['ur_code'])},404
         
-        order.name = data['order_name']
-        order.staff_id = data['staff_id']
-        order.user_id = data['user_id']
-        order.save_to_db()
-        return {"message":"order info updated succesfully."},200
+        # only admin and staff(post owner) are allowed to modify existing orders.
+        identity = get_jwt_identity()
+
+        if identity["auth_level"]=="admin" or (
+            identity["auth_level"] == "staff" and identity["id"] == order.staff_id): 
+
+            order.name = data['order_name']
+            order.staff_id = data['staff_id']
+            order.user_id = data['user_id']
+            order.save_to_db()
+            return {"message":"order info updated succesfully."},200
+        
+        else:
+            return {"message":"unauthorized access for modififying order."},500
 
 #################################################
 #### soft delete order ########################
@@ -117,15 +131,24 @@ class OrderDelete(Resource):
     order_parser.add_argument(
         "ur_code", type = str, required=True, help="ur_code cannot be blank."
     )
-    # @jwt_required
+    @jwt_required
     def delete(self):
         data = self.order_parser.parse_args()
 
         order = OrderModel.find_by_ur_code(data['ur_code'])
-
         if not order:
             return {"message":"order with ur_code{} doen not exist.".format(data['ur_code'])},404
         
-        order.delete_from_db()
+        # only admin and staff(post owner) are allowed to modify existing orders.
+        identity = get_jwt_identity()
 
-        return {"message":"order deleted succesfully."},200
+        if identity["auth_level"]=="admin" or (
+            identity["auth_level"] == "staff" and identity["id"] == order.staff_id): 
+            try:
+                order.delete_from_db()
+                return {"message":"order deleted succesfully."},200
+            except:
+                return{"message":"something went wrong."}
+
+        else:
+            return {"message":"unauthorized access for modififying order."},500
